rename modules to snake case #9
