+++ src/client.rs
@@ -2,5 +2,5 @@ pub use crate::{
     cli::{client as cli, client::Cli},
     config::{client as config, client::Config},
     defaults::client as defaults,
-    keymgmt::client as keymgmt,
+    key_mgmt::client as keymgmt,
 };
