+++ src/server.rs
@@ -2,6 +2,6 @@ pub use crate::{
     cli::{server as cli, server::Cli},
     config::{server as config, server::Config},
     defaults::server as defaults,
-    keymgmt::server as keymgmt,
+    key_mgmt::server as keymgmt,
 };
 pub use transport::server::{self as server, Chan, Server};
